Safari-609.1.11.1/LayoutTests/editing/pasteboard/merge-end-1.html

<script src="../../../resources/dump-as-markup.js"></script>
<!DOCTYPE html>
<html>
<body>
<div id="test" contenteditable="true"><span>foo</span></div>

<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
<script>
var s = window.getSelection();
var e = document.getElementById("test");

e.focus();
moveSelectionForwardByCharacterCommand();
insertHTMLCommand("<div>bar</div>baz");

Markup.description('This tests the last paragraph in the pasted fragment is merged with the content after the insertion position.\n'
+ '"baz" and "oo" should be in the same paragraph below.');
Markup.dump('test');

</script>
</body>
</html>
